practical_astronomy_rust

Module coordinates

Functions

angle_between_two_objects

Calculate the angle between two celestial objects.

angle_to_decimal_degrees

Convert an Angle (degrees, minutes, and seconds) to Decimal Degrees.

atmospheric_refraction

Calculate corrected RA/Dec, accounting for atmospheric refraction.

carrington_rotation_number

Calculate carrington rotation number for a Greenwich date.

correct_for_aberration

Correct ecliptic coordinates for the effects of aberration.

correct_for_precession

Calculate precession (corrected coordinates between two epochs).

corrections_for_geocentric_parallax

Calculate corrected RA/Dec, accounting for geocentric parallax.

decimal_degrees_to_angle

Convert Decimal Degrees to an Angle (degrees, minutes, and seconds).

ecliptic_coordinate_to_equatorial_coordinate

Convert Ecliptic Coordinates to Equatorial Coordinates.

equatorial_coordinate_to_ecliptic_coordinate

Convert Equatorial Coordinates to Ecliptic Coordinates.

equatorial_coordinate_to_galactic_coordinate

Convert Equatorial Coordinates to Galactic Coordinates.

equatorial_coordinates_to_horizon_coordinates

Convert Equatorial Coordinates to Horizon Coordinates.

galactic_coordinate_to_equatorial_coordinate

Convert Galactic Coordinates to Equatorial Coordinates.

heliographic_coordinates

Calculate heliographic coordinates for a given Greenwich date, with a given heliographic position angle and heliographic displacement in arc minutes.

horizon_coordinates_to_equatorial_coordinates

Convert Horizon Coordinates to Equatorial Coordinates.

hour_angle_to_right_ascension

Convert Hour Angle to Right Ascension.

mean_obliquity_of_the_ecliptic

Calculate Mean Obliquity of the Ecliptic for a Greenwich Date.

nutation_in_ecliptic_longitude_and_obliquity

Calculate nutation for two values: ecliptic longitude and obliquity, for a Greenwich date.

right_ascension_to_hour_angle

Convert Right Ascension to Hour Angle.

rising_and_setting

Rising and Setting times.

selenographic_coordinates_1

Calculate selenographic (lunar) coordinates (sub-Earth).

selenographic_coordinates_2

Calculate selenographic (lunar) coordinates (sub-Solar).