Module practical_astronomy_rust::sun

Functions

approximate_position_of_sun

Calculate approximate position of the sun for a local date and time.

equation_of_time

Calculate the equation of time. (The difference between the real Sun time and the mean Sun time.)

morning_and_evening_twilight

Calculate times of morning and evening twilight.

precise_position_of_sun

Calculate precise position of the sun for a local date and time.

solar_elongation

Calculate solar elongation for a celestial body.

sun_distance_and_angular_size

Calculate distance to the Sun (in km), and angular size.

sunrise_and_sunset

Calculate local sunrise and sunset.