practical_astronomy_rust

Module util

Functions

bool_to_int

Given a boolean value, return a corresponding int (1 or 0)

get_local_hour_from_ut

Convert a Universal Time hour to Local Time

is_leap_year

Determine if year is a leap year.

round_f64

Round an f64 primitive to the specified number of decimal places.

ternary_assign

Return one of two possible values, dependent upon the value of is_true_state