Module practical_astronomy_rust::macros

Structs

PlDataStruct

Functions

angle

Calculate the angle between two celestial objects

cd_jd

Convert a Greenwich Date/Civil Date (day,month,year) to Julian Date

dd_deg

Return Degrees part of Decimal Degrees

dd_dh

Convert Decimal Degrees to Degree-Hours

dd_min

Return Minutes part of Decimal Degrees

dd_sec

Return Seconds part of Decimal Degrees

degrees

Convert W to Degrees

dh_dd

Convert Degree-Hours to Decimal Degrees

dh_hour

Return the hour part of a Decimal Hours

dh_min

Return the minutes part of a Decimal Hours

dh_sec

Return the seconds part of a Decimal Hours

dms_dd

Convert Degrees Minutes Seconds to Decimal Degrees

e_gst_ut

Status of conversion of Greenwich Sidereal Time to Universal Time.

e_moon_rise

Moonrise calculation status.

e_moon_rise_l6680

Helper function for e_moon_rise()

e_moon_rise_l6700

Helper function for e_moon_rise()

e_moon_set

Moonset calculation status.

e_moon_set_l6680

Helper function for e_moon_set()

e_moon_set_l6700

Helper function for e_moon_set()

e_rs

Rise/Set status

e_sun_rs

Sunrise/Sunset calculation status.

e_sun_rs_l3710

Helper function for e_sun_rs()

e_twilight

Twilight calculation status.

e_twilight_l3710

Helper function for e_twilight()

ec_dec

Ecliptic - Declination (degrees)

ec_ra

Ecliptic - Right Ascension (degrees)

eccentric_anomaly

Solve Kepler’s equation, and return value of the eccentric anomaly in radians

eq_alt

Convert Equatorial Coordinates to Altitude (in decimal degrees)

eq_az

Convert Equatorial Coordinates to Azimuth (in decimal degrees)

eq_e_lat

Original macro name: EQElat

eq_e_long

Original macro name: EQElong

f_dow

Convert a Julian Date to Day-of-Week (e.g., Sunday)

f_part

Original macro name: Fpart

fract

Original macro name: FRACT

full_moon

Calculate Julian date of Full Moon.

gst_lst

Convert Greenwich Sidereal Time to Local Sidereal Time

gst_ut

Convert Greenwich Sidereal Time to Universal Time

ha_ra

Convert Hour Angle to Right Ascension

hms_dh

Convert a Civil Time (hours,minutes,seconds) to Decimal Hours

hor_dec

Convert Horizon Coordinates to Declination (in decimal degrees)

hor_ha

Convert Horizon Coordinates to Hour Angle (in decimal degrees)

iint

Original macro name: IINT

jdc_day

Returns the day part of a Julian Date

jdc_month

Returns the month part of a Julian Date

jdc_year

Returns the year part of a Julian Date

lct_gday

Determine Greenwich Day for Local Time

lct_gmonth

Determine Greenwich Month for Local Time

lct_gyear

Determine Greenwich Year for Local Time

lct_ut

Convert Local Civil Time to Universal Time

lint

Original macro name: LINT

lst_gst

Convert Local Sidereal Time to Greenwich Sidereal Time

lunar_eclipse_occurrence

Determine if a lunar eclipse is likely to occur.

lunar_eclipse_occurrence_l6855

Helper function for lunar_eclipse_occurrence

mag_lunar_eclipse

Calculate magnitude of lunar eclipse.

mag_solar_eclipse

Calculate magnitude of solar eclipse.

mag_solar_eclipse_l7390

Helper function for mag_solar_eclipse

moon_dist

Calculate distance from the Earth to the Moon (km)

moon_hp

Calculate horizontal parallax for the Moon

moon_lat

Calculate geocentric ecliptic latitude for the Moon

moon_long

Calculate geocentric ecliptic longitude for the Moon

moon_long_lat_hp

Calculate longitude, latitude, and horizontal parallax of the Moon.

moon_mean_anomaly

Calculate the Moon’s mean anomaly.

moon_phase

Calculate current phase of Moon.

moon_rise_az

Local azimuth of moonrise.

moon_rise_az_l6680

Helper function for moon_rise_az

moon_rise_az_l6700

Helper function for moon_rise_az

moon_rise_lc_dmy

Local date of moonrise.

moon_rise_lc_dmy_l6680

Helper function for moon_rise_lc_dmy

moon_rise_lc_dmy_l6700

Helper function for moon_rise_lc_dmy

moon_rise_lct

Local time of moonrise.

moon_rise_lct_l6680

Helper function for moon_rise_lct

moon_rise_lct_l6700

Helper function for moon_rise_lct

moon_set_az

Local azimuth of moonset.

moon_set_az_l6680

Helper function for moon_set_az

moon_set_az_l6700

Helper function for moon_set_az

moon_set_lc_dmy

Local date of moonset.

moon_set_lc_dmy_l6680

Helper function for moon_set_lc_dmy

moon_set_lc_dmy_l6700

Helper function for moon_set_lc_dmy

moon_set_lct

Local time of moonset.

moon_set_lct_l6680

Helper function for moon_set_lct

moon_set_lct_l6700

Helper function for moon_set_lct

moon_size

Calculate the Moon’s angular diameter (degrees)

new_moon

Calculate Julian date of New Moon.

new_moon_full_moon_l6855

Helper function for new_moon() and full_moon() “”“

nutat_long

Nutation amount to be added in ecliptic longitude, in degrees.

nutat_obl

Nutation of Obliquity

obliq

Obliquity of the Ecliptic for a Greenwich Date

p_comet_long_lat_dist

Calculate longitude, latitude, and distance of parabolic-orbit comet.

parallax_dec

Calculate corrected declination in decimal degrees

parallax_dec_l2870

Helper function for parallax_dec

parallax_ha

Calculate corrected hour angle in decimal hours

parallax_ha_l2870

Helper function for parallax_ha

planet_coordinates

Calculate several planetary properties.

planet_long_l4685

Helper function for planet_long_lat()

planet_long_l4735

Helper function for planet_long_lat()

planet_long_l4810

Helper function for planet_long_lat()

planet_long_l4945

Helper function for planet_long_lat()

ra_ha

Convert Right Ascension to Hour Angle

refract

Calculate effects of refraction

refract_l3035

Helper function for refract

rise_set_azimuth_rise

Azimuth of rising, in degrees.

rise_set_azimuth_set

Azimuth of setting, in degrees.

rise_set_local_sidereal_time_rise

Local sidereal time of rise, in hours.

rise_set_local_sidereal_time_set

Local sidereal time of setting, in hours.

sgn

Calculate sign of number.

solar_eclipse_occurrence

Determine if a solar eclipse is likely to occur.

solar_eclipse_occurrence_l6855

Helper function for solar_eclipse_occurrence

solve_cubic

For W, in radians, return S, also in radians.

sun_dia

Calculate Sun’s angular diameter in decimal degrees

sun_dist

Calculate Sun’s distance from the Earth in astronomical units

sun_e_long

Mean ecliptic longitude of the Sun at the epoch

sun_ecc

Eccentricity of the Sun-Earth orbit

sun_long

Calculate Sun’s ecliptic longitude

sun_mean_anomaly

Calculate the Sun’s mean anomaly.

sun_peri

Longitude of the Sun at perigee

sun_true_anomaly

Calculate Sun’s true anomaly, i.e., how much its orbit deviates from a true circle to an ellipse.

sunrise_az

Calculate azimuth of sunrise.

sunrise_az_l3710

Helper function for sunrise_az()

sunrise_lct

Calculate local civil time of sunrise.

sunrise_lct_l3710

Helper function for sunrise_lct()

sunset_az

Calculate azimuth of sunset.

sunset_az_l3710

Helper function for sunset_az()

sunset_lct

Calculate local civil time of sunset.

sunset_lct_l3710

Helper function for sunset_lct().

true_anomaly

Solve Kepler’s equation, and return value of the true anomaly in radians

twilight_am_lct

Calculate morning twilight start, in local time.

twilight_am_lct_l3710

Helper function for twilight_am_lct()

twilight_pm_lct

Calculate evening twilight end, in local time.

twilight_pm_lct_l3710

Helper function for twilight_pm_lct()

unwind

Convert angle in radians to equivalent angle in degrees.

unwind_deg

Convert angle in degrees to equivalent angle in the range 0 to 360 degrees.

unwind_rad

Convert angle in radians to equivalent angle in degrees.

ut_day_adjust

Original macro name: UTDayAdjust

ut_end_total_lunar_eclipse

Calculate end time of total phase of lunar eclipse (UT)

ut_end_umbra_lunar_eclipse

Calculate end time of umbra phase of lunar eclipse (UT)

ut_first_contact_lunar_eclipse

Calculate time of first shadow contact for lunar eclipse (UT)

ut_first_contact_solar_eclipse

Calculate time of first contact for solar eclipse (UT)

ut_first_contact_solar_eclipse_l7390

Helper function for ut_first_contact_solar_eclipse

ut_gst

Convert Universal Time to Greenwich Sidereal Time

ut_last_contact_lunar_eclipse

Calculate time of last shadow contact for lunar eclipse (UT)

ut_last_contact_solar_eclipse

Calculate time of last contact for solar eclipse (UT)

ut_last_contact_solar_eclipse_l7390

Helper function for ut_last_contact_solar_eclipse

ut_lc_day

Get Local Civil Day for Universal Time

ut_lc_month

Get Local Civil Month for Universal Time

ut_lc_year

Get Local Civil Year for Universal Time

ut_lct

Convert Universal Time to Local Civil Time

ut_max_lunar_eclipse

Calculate time of maximum shadow for lunar eclipse (UT)

ut_max_solar_eclipse

Calculate time of maximum shadow for solar eclipse (UT)

ut_max_solar_eclipse_l7390

Helper function for ut_max_solar_eclipse

ut_start_total_lunar_eclipse

Calculate start time of total phase of lunar eclipse (UT)

ut_start_umbra_lunar_eclipse

Calculate start time of umbra phase of lunar eclipse (UT)